app_cart: Move class Cart into cart.py
